backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
groq>=0.9.0
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
gitpython>=3.1.43
tree-sitter>=0.22.0
tree-sitter-language-pack>=0.9.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
